Backend/requirements.txt
fastapi
uvicorn
langchain
pymilvus
openai
langgraph
python-dotenv
langchain-openai
datetime
langchain-milvus
datetime
pydantic
uuid
requests
twilio
logging
asyncio
gunicorn
langchain_google_genai